правки

+++ ru.ilya.SpringMvc/src/main/java/DAO/UserDAO.java
@@ -9,5 +9,5 @@ public interface UserDAO {
     List<User> allUser();
     void add(User user);
     void edit(User user);
-    User getById(int id);
+    User getById(long id);
 }
